requirements.txt
fastapi
uvicorn
PyPDF2
httpx
pydantic
python-dotenv